% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqiAnnotation.R
\name{aqiAnnotation}
\alias{aqiAnnotation}
\title{Add AQI annotations an existing ggplot}
\usage{
aqiAnnotation(gg, xlo, xhi, ylo, yhi, style = "bars_lines",
  barWidth = 0.01, lineSize = 0.5)
}
\arguments{
\item{gg}{\code{ggplot} object.}

\item{xlo}{Left edge of the plot as a \code{POSIXct} value.}

\item{xhi}{Right edge of the plot as a \code{POSIXct} value.}

\item{ylo}{Bottom edge of the plot as a \code{numeric} value.}

\item{yhi}{Top edge of the plot as a \code{numeric} value.}

\item{style}{Annotation style as described below.}

\item{barWidth}{Width of stacked bars as a percentage of the X axis.}

\item{lineSize}{Size of horizontal lines.}
}
\value{
A `ggplot` plot object with AQI annotations.
}
\description{
Adds AQI annotations to an existing \code{ggplot}. The \code{style} parameter
must contain one ore more of the following options separated by underscores:

\itemize{
\item{\code{bars} -- stacked color bars}
\item{\code{lines} -- colored lines}
\item{\code{labels} -- AQI labels}
}

The returned object may be further amended with `ggplot` elements.
}
