% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockPlotBase.R
\name{clockPlotBase}
\alias{clockPlotBase}
\title{Create a Clock plot for a single monitor}
\usage{
clockPlotBase(ws_monitor, startdate = NULL, enddate = NULL,
  title = "", centerColor = "black", gapColor = "black",
  gapFraction = 1/25, plotRadius = 1, dataRadii = c(0.5, 1),
  shadedNight = FALSE, hoursPerTick = NULL, solarLabels = FALSE,
  colorPalette = aqiPalette("aqi"))
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object containing a single monitor.}

\item{startdate}{Desired start date (integer or character in Ymd format 
or \code{POSIXct}).}

\item{enddate}{Desired end date (integer or character in Ymd format
or \code{POSIXct}).}

\item{title}{Optional title for the plot.}

\item{centerColor}{Color used for the center of the circle.}

\item{gapColor}{Color used for the day-break gap.}

\item{gapFraction}{Fraction of the circle used as the day boundary gap.}

\item{plotRadius}{Full radius of the plot.}

\item{dataRadii}{Inner and outer radii for the data portion of the plot [0:1].}

\item{shadedNight}{Add nighttime shading.}

\item{hoursPerTick}{Add tick marks every # hours. Defaults to no ticks.}

\item{solarLabels}{Add sunrise/sunset labels.}

\item{colorPalette}{Palette function to convert monitor values into colors.

Room for annotations can be created by setting \code{plotRadius = 1.2}.

TODO:  More documentation}
}
\value{
A **ggplot** plot object with a "clock plot" for a single monitor.
}
\description{
Create a "clock plot" showing PM2.5 data for a single day for the given 
monitors. A colored bar curves around in a clockwise manner with 12/4 of the
bar colored for each hour of the local time day.
}
\examples{
ws_monitor <- PWFSLSmoke::Carmel_Valley
startdate <- "2016-08-07"
clockPlotBase(ws_monitor, startdate)
}
