% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarnayPlot.R
\name{createTarnayPlot}
\alias{createTarnayPlot}
\title{Create a Tarnay plot for many monitors}
\usage{
createTarnayPlot(monitors, data, columns = 1, title = NULL, xLabel = NULL,
  yLabel = NULL)
}
\arguments{
\item{monitors}{Monitor ID(s) to create plot for.}

\item{data}{Data used to create plot (NOTE: currently must be a `ws_monitor`
object).}

\item{columns}{Number of columns the faceted plot should have (default 1).}

\item{title}{The title of the plot. Defaults to specifying the types of
data present in the plot.}

\item{xLabel}{The x-axis label of the plot. Defaults to years present in
data.}

\item{yLabel}{The y-axis label of the plot. Defaults to PM2.5.}
}
\value{
A **ggplot** plot of the given monitors and data.
}
\description{
Create a timeseries barplot showing PM2.5 data for the given monitors. The
overall plot is faceted by monitor, and each facet has two sets of columns:
one for daily levels, and one for hourly levels.
}
\examples{

}
