% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_aqiAnnotation.R
\name{custom_aqiLines}
\alias{custom_aqiLines}
\title{Add AQI lines to a plot}
\usage{
custom_aqiLines(...)
}
\arguments{
\item{...}{Arguments passed on to layer, such as aesthetic properties like
size or alpha.}
}
\value{
A `ggplot` plot object with AQI annotations.
}
\description{
Adds AQI lines to a plot
}
