% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_aqiAnnotation.R
\name{custom_aqiStackedBar}
\alias{custom_aqiStackedBar}
\title{Add AQI stacked bars to a plot}
\usage{
custom_aqiStackedBar(width = 0.02, position = "identity", ...)
}
\arguments{
\item{width}{Width of bars as a fraction of plot width.}

\item{position}{Position adjustment, either as a string, or the result of 
a call to a position adjustment function}

\item{...}{additional arguments passed on to layer, such as alpha.}
}
\value{
A `ggplot` plot object with AQI annotations.
}
\description{
Adds AQI stacked bars to a plot.
}
