% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_currentNowcast.R
\name{custom_currentNowcast}
\alias{custom_currentNowcast}
\title{Add a bar for the current NowCast to a plot}
\usage{
custom_currentNowcast(ws_tidy, monitorID = NULL, timezone = NULL,
  width = 0.8)
}
\arguments{
\item{monitorID}{}
}
\description{
Add a bar for the current NowCast to a plot
}
