% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_datetimeScale.R
\name{custom_datetimeScale}
\alias{custom_datetimeScale}
\title{Add date scale with custom styling for timezone and daterange}
\usage{
custom_datetimeScale(startdate = NULL, enddate = NULL,
  timezone = NULL, expand = c(0, 0.05), break_width = NULL,
  minor_break_width = NULL, date_labels = "\%b \%d",
  tick_location = c("midnight", "midday")[1],
  includeFullEnddate = TRUE, ...)
}
\arguments{
\item{startdate}{Desired axis start date limit, in a format that can be 
parsed with \link{parseDatetime}.}

\item{enddate}{Desired axis end date limit, in a format that can be 
parsed with \link{parseDatetime}.}

\item{timezone}{Timezone for label formatting.}

\item{expand}{Vector of range expansion constants used to add some padding 
around the data, to ensure that they are placed some distance away from the 
axes.}

\item{break_width}{Space between breaks. If \code{NULL}, suitable breaks are
calculated based on the data. See 'Details'.}

\item{minor_break_width}{Space between minor breaks. If \code{NULL}, suitable
breaks are calculated based on the data. See 'Details'.}

\item{date_labels}{date format string for formatting date labels.}

\item{tick_location}{Location of ticks ("midnight" or "midday")}

\item{includeFullEnddate}{if \code{TRUE}, the x-axis limit is pushed up to include
the full final day.}

\item{...}{Additional arguments passed onto \code{\link[ggplot2]{scale_x_datetime}}.}
}
\description{
Add a date scale and custom formatting for creating 
consistent timeseries plots.
}
