% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_legend.R
\name{custom_legend}
\alias{custom_legend}
\title{Add a custom legend to a plot}
\usage{
custom_legend(title = "", labels = NULL, aesthetics = NULL,
  theme_args = list(legend.position = "right", legend.direction =
  "vertical"), ...)
}
\arguments{
\item{title}{Legend title.}

\item{labels}{Labels for legend elements}

\item{aesthetics}{List of aesthetic for legend elements. Each 
list item should have the name of the aesthetic (like \code{linetype} or \code{color}), 
and the value should be a vector the same length as \code{labels}.}

\item{theme_args}{List of arguments to be passed on to \code{theme}, such as 
\code{legend.position}.}

\item{...}{Additional arguments passed on to \code{guide_legend()}.}
}
\description{
Add a default legend to a plot.
}
