% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyBarplot.R
\name{dailyBarplot}
\alias{dailyBarplot}
\title{Create a daily barplot for a single monitor}
\usage{
dailyBarplot(ws_monitor, startdate = NULL, enddate = NULL,
  style = NULL, monitorID = NULL, currentNowcast = NULL,
  currentPrediction = NULL, title = "")
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object.}

\item{startdate}{Desired start date (integer or character in Ymd format
or \code{POSIXct}).}

\item{enddate}{Desired end date (integer or character in Ymd format
or \code{POSIXct}).}

\item{style}{Plot style.}

\item{monitorID}{Monitor ID of interest. Required if \code{ws_monitor} 
contains more than one monitor.}

\item{currentNowcast}{Real-time current Nowcast value -- for use in plots 
presented in the PWFSL monitoring site.}

\item{currentPrediction}{Real-time current prediction for today's daily 
average -- for use in plots presented in the PWFSL monitoring site.}

\item{title}{Optional title for the plot.}
}
\value{
A `ggplot` plot object with a "daily barplot" for a single monitor.
}
\description{
Create a daily average barplot of PM2.5 values at a location.

This function presents a simplified interface to \code{\link{dailyBarplotBase}}
and collects common options into a set of named styles.

Current styles include:
\itemize{
\item{\code{pwfsl} -- used in the PWFSL monitoring site. Best for 1-2
weeks of data.}
}

The returned object may be further amended with **ggplot** elements.
}
\details{
Daily averages are calculated from midnight-to-midnight in monitor 
local time.
}
\examples{
NW <- PWFSLSmoke::Northwest_Megafires
dailyBarplot(NW, "2015-07-01", "2015-10-01",
             style = "month",
             monitorID = "160690014_01",
             title = "Daily Average PM2.5\\nRuebens, Idaho")
}
\seealso{
\code{\link{dailyBarplotBase}}
}
