% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{.nowcast}
\alias{.nowcast}
\title{Calculate NowCast}
\usage{
.nowcast(x, version = "pm", includeShortTerm = FALSE)
}
\arguments{
\item{x}{Vector of ordered data for the Nowcast algorithm to be applied to.}

\item{version}{character identity specifying the type of nowcast algorithm to be used. 
For details see \link{monitor_nowcast}.}

\item{includeShortTerm}{calculate preliminary NowCast values starting with the 2nd hour.}
}
\description{
This function calculates the NowCast values for the data, and adds  it to a plot.
The default is to add a NowCast line.
}
