% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{isWSMon}
\alias{isWSMon}
\title{Check if Data has `ws_monitor` Format}
\usage{
isWSMon(data = NULL)
}
\arguments{
\item{data}{Data to validate.}
}
\value{
True if the data is of class `ws_monitor`, otherwise False.
}
\description{
Verifies that the given data can be treated as a `ws_monitor`
    object.
}
