% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_legend_pm25Timeseries.R
\name{legend_pm25Timeseries}
\alias{legend_pm25Timeseries}
\title{Add pm25Timeseries legend to pm25Timeseries plot}
\usage{
legend_pm25Timeseries(legend.labels = c("PM2.5", "NowCast"),
  legend.title = "", color = c("black", "black"), linetype = c(NA,
  1), shape = c(16, NA))
}
\arguments{
\item{legend.labels}{Labels for the legend. Must match names for color aesthetics.}

\item{legend.title}{Title for the legend}

\item{color}{Vector of colors. This also affects the objects which the legend 
represents.}

\item{linetype}{Vector of linetypes for display.}

\item{shape}{Vector of shapes for display.}
}
\description{
Add a default legend to a timeseries plot. geom_pm25Points and stat_nowcast are
designed to work with this function, to create a tidy_timeseries plot.
}
