% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_ggClockPlot.R
\name{monitor_ggClockPlot}
\alias{monitor_ggClockPlot}
\title{Create clock plot for one monitor}
\usage{
monitor_ggClockPlot(ws_monitor, startdate = NULL, enddate = NULL,
  monitorID = NULL, timezone = NULL, ...)
}
\arguments{
\item{ws_monitor}{A `ws_monitor` object}

\item{startdate}{Desired startdate for data to include, in a format that can be 
parsed with \link{parseDatetime}.}

\item{enddate}{Desired enddate for data to include, in a format that can be parsed 
with \link{parseDatetime}.}

\item{monitorID}{monitorID to include in the plot.}

\item{timezone}{Timezone for x-axis scale. If NULL and only one timezone present
in the data, the data timezone will be used. If NULL and multiple timezones 
present, the default is UTC.}

\item{...}{Additional arguments passed on to \code{\link{custom_pm25DiurnalScales}}.}
}
\value{
A **ggplot** object
}
\description{
This function wraps \link{tidy_ggDailyBarplot}, accepting a `ws_monitor` 
object to assemble various layers to create a production-ready
timeseries plot for one or more monitors.
}
\examples{
\dontrun{
ws_monitor <- airnow_loadLatest()
monitor_ggClockPlot(ws_monitor, monitorID = "410432002_01")
}

ws_monitor <- Carmel_Valley
monitor_ggClockPlot(ws_monitor, 
                    startdate = 20160801,
                    enddate = 20160810)

}
