% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_ggDailyBarplot_archival.R
\name{monitor_ggDailyBarplot_archival}
\alias{monitor_ggDailyBarplot_archival}
\title{Create an archival daily barplot for one or more monitors}
\usage{
monitor_ggDailyBarplot_archival(
  ws_monitor,
  startdate = NULL,
  enddate = NULL,
  monitorID = NULL,
  style = "large",
  title = NULL,
  timezone = NULL,
  today = TRUE,
  ...
)
}
\arguments{
\item{ws_monitor}{A \code{ws_monitor} object.}

\item{startdate}{Desired start date (integer or character in ymd format or
POSIXct).}

\item{enddate}{Desired end date (integer or character in ymd format or
POSIXct).}

\item{monitorID}{monitorID to include in the plot. This can be NULL if
\emph{ws_monitor} only has one unique monitorID.}

\item{style}{String indicating plotting style. Either \code{"large"} or
\code{"small"}. \code{style = "large"} is suitable for plots larger than
450x450px, and \code{"small"} is suitable for plots 450x450px or smaller.}

\item{title}{Plot title. If NULL, a suitable title will be constructed.}

\item{timezone}{Olson timezone name for x-axis scale and date parsing. If
NULL the timezone of the specified monitor will be used.}

\item{today}{Logical indicating whether to include a shaded "current NowCast"
bar for Today. Ignored if data is not current.}

\item{...}{Arguments passed onto \code{\link{ggplot_pm25Timeseries}}.}
}
\value{
A \emph{ggplot} object
}
\description{
This function assembles various layers to create a production-ready
archival daily barplot for one or more monitors.

The full range of data in \code{ws_monitor} will be used unless both
\code{startdate} and \code{enddate} are specified.
}
\examples{
library(AirMonitorPlots)

PWFSLSmoke::Carmel_Valley \%>\%
  monitor_trimDate() \%>\%
  monitor_ggDailyBarplot_archival()

\dontrun{
ws_monitor <- airnow_loadLatest()
monitor_ggDailyBarplot_archival(
  ws_monitor,
  monitorID = "410432002_01",
  today = TRUE
)
}
}
