% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_ggDailyHourlyBarplot.R
\name{monitor_ggDailyHourlyBarplot}
\alias{monitor_ggDailyHourlyBarplot}
\title{Create a Daily-Hourly plot for many monitors}
\usage{
monitor_ggDailyHourlyBarplot(
  monitor,
  startdate = NULL,
  enddate = NULL,
  id = NULL,
  columns = 1,
  title = NULL,
  timezone = NULL,
  xLabel = NULL,
  yLabel = NULL,
  hourlyDataType = c("nowcast", "raw", "none"),
  hourlyBarWidth = 0.6,
  palette = "epa_aqi",
  includeLegend = FALSE
)
}
\arguments{
\item{monitor}{A \emph{mts_monitor} object.}

\item{startdate}{Desired start date (integer or character in ymd format or
POSIXct).}

\item{enddate}{Desired end date (integer or character in ymd format or
POSIXct).}

\item{id}{Optional vector of deviceDeploymentIDs used to filter the data.}

\item{columns}{Number of columns the faceted plot should have.}

\item{title}{The title of the plot. Defaults to specifying the types of
data present in the plot.}

\item{timezone}{Olson timezone name for x-axis scale and date parsing. If
NULL the timezone of the specified monitor will be used.}

\item{xLabel}{The x-axis label of the plot. Defaults to years present in
data.}

\item{yLabel}{The y-axis label of the plot. Defaults to PM2.5.}

\item{hourlyDataType}{The type of hourly data to include in the plot. The
options include "nowcast" (hourly nowcast values), "raw" (raw hourly values),
or "none" (no hourly data at all).}

\item{hourlyBarWidth}{Fractional width of hourly bars relative to one hour.
Defaults to 0.45. Values near 1 produce wide bars with little spacing,
while smaller values produce narrower bars with more separation.}

\item{palette}{The ordered color palette used to represent each AQI
category. Currently defaults to (and only accepts) "epa_aqi".}

\item{includeLegend}{Option to include a legend..}
}
\value{
A \strong{ggplot} plot of the given monitors and data.
}
\description{
Create a time series barplot showing PM2.5 data for the given monitors. The
overall plot is faceted by monitor, and each facet has two sets of columns:
one for daily levels, and one for hourly levels.

The full range of data in \code{monitor} will be used unless both
\code{startdate} and \code{enddate} are specified.

The timezone specified or, if \code{timezone = NULL}, that of the first
monitor encountered will be used for all time axes.
}
\examples{
\dontrun{
library(AirMonitor)
library(AirMonitorPlots)

# Fail gracefully if any resources are not available
try({

now <- lubridate::now('America/Los_Angeles')
start <- now - lubridate::ddays(6)

SF <-
  monitor_loadLatest() \%>\%
  monitor_filterByDistance(
    longitude = -122.42,
    latitude = 37.77,
    radius = 20000,
    count = 4
  ) \%>\%
  monitor_filterDate(
    startdate = start,
    enddate = now
  )

monitor_ggDailyHourlyBarplot(SF)

}, silent = FALSE)
}
}
