% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_ggTimeseries.R
\name{monitor_ggTimeseries}
\alias{monitor_ggTimeseries}
\title{Create a timeseries plot for one or more monitors}
\usage{
monitor_ggTimeseries(
  mts_monitor,
  startdate = NULL,
  enddate = NULL,
  deviceDeploymentIDs = NULL,
  style = "small",
  title = NULL,
  aqiStyle = NULL,
  timezone = NULL,
  ...
)
}
\arguments{
\item{mts_monitor}{A \code{mts_monitor} object.}

\item{startdate}{Desired start date (integer or character in ymd format or
POSIXct).}

\item{enddate}{Desired end date (integer or character in ymd format or
POSIXct).}

\item{deviceDeploymentIDs}{vector of deviceDeploymentIDs to include in the plot. If more than
one, different monitors will be plotted in different colors.}

\item{style}{Plot style. \code{small} or \code{large}. \code{style = small}
is appropriate for plots 450x450px or smaller; \code{style = large} is
appropriate for plots larger than 450x450px.}

\item{title}{Plot title. If NULL, a suitable title will be constructed.}

\item{aqiStyle}{AQI style to add AQI color bars, lines, and labels. Not
currently supported.}

\item{timezone}{Olson timezone name for x-axis scale and date parsing. If
NULL the timezone of the specified monitor will be used.}

\item{...}{Arguments passed onto \code{\link{ggplot_pm25Timeseries}}.}
}
\value{
A \emph{ggplot} object.
}
\description{
This function assembles various layers to create a production-ready
timeseries plot for one or more monitors.
}
\examples{
NM <- AirMonitor::Northwest_Megafires
monitor_ggTimeseries(
  NM,
  startdate = 20150815,
  enddate = 20150831,
  deviceDeploymentIDs = "160690014_01"
)

CV <- AirMonitor::Carmel_Valley
monitor_ggTimeseries(CV, startdate = 20160801, enddate = 20160810)

\dontrun{
mts_monitor <- airnow_loadLatest()
monitor_ggTimeseries(mts_monitor, deviceDeploymentID = "410432002_01")
}
}
