% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_ggTimeseries_archival.R
\name{monitor_ggTimeseries_archival}
\alias{monitor_ggTimeseries_archival}
\title{Create an archival timeseries plot for one or more monitors}
\usage{
monitor_ggTimeseries_archival(
  monitor,
  startdate = NULL,
  enddate = NULL,
  id = NULL,
  style = "large",
  title = NULL,
  timezone = NULL,
  ...
)
}
\arguments{
\item{monitor}{A \emph{mts_monitor} object.}

\item{startdate}{Desired start date (integer or character in ymd format or
POSIXct).}

\item{enddate}{Desired end date (integer or character in ymd format or
POSIXct).}

\item{id}{vector of deviceDeploymentIDs to include in the plot. If more than
one, different monitors will be plotted in different colors.}

\item{style}{Plot style. \code{small} or \code{large}. \code{style = small}
is appropriate for plots 450x450px or smaller; \code{style = large} is
appropriate for plots larger than 450x450px.}

\item{title}{Plot title. If NULL, a suitable title will be constructed.}

\item{timezone}{Olson timezone name for x-axis scale and date parsing. If
NULL the timezone of the specified monitor will be used.}

\item{...}{Arguments passed onto \code{\link{ggplot_pm25Timeseries}}.}
}
\value{
A \emph{ggplot} object.
}
\description{
This function assembles various layers to create a production-ready
archival timeseries plot for one or more monitors.
}
\examples{
library(AirMonitorPlots)

AirMonitor::NW_Megafires \%>\%
  monitor_ggTimeseries_archival(
    startdate = 20150809,
    enddate = 20150820,
    id = "575243c65b9e4719_160690012",
    timezone = "America/Los_Angeles"
 )

AirMonitor::Carmel_Valley \%>\%
  AirMonitor::monitor_trimDate() \%>\%
  monitor_ggTimeseries_archival()

}
