% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_trimDate.R
\name{monitor_trimDate}
\alias{monitor_trimDate}
\title{Trim a ws_monitor object to full days}
\usage{
monitor_trimDate(ws_monitor = NULL, timezone = NULL)
}
\arguments{
\item{ws_monitor}{ws_monitor \emph{ws_monitor} object.}

\item{timezone}{Olson timezone used to interpret dates.}
}
\value{
A subset of the given \emph{ws_monitor} object.
}
\description{
Trims the date range of a \emph{ws_monitor} object to local time date
boundaries which are \emph{within} the range of data. This has the effect
of removing partial-day data records at the start and end of the timeseries
and is useful when calculating full-day statistics.

Day boundaries are calculated using the specified \code{timezone} or, if
\code{NULL}, from \code{ws_monitor$meta$timezone}.
}
\examples{
library(AirMonitorPlots)

UTC_week <- PWFSLSmoke::monitor_subset(
  PWFSLSmoke::Carmel_Valley,
  tlim = c(20160801, 20160808),
  timezone = "UTC"
)

# UTC day boundaries
head(UTC_week$data)

# Trim to local time day boundaries
local_week <- monitor_trimDate(UTC_week)
head(local_week$data)

}
