% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_pwfsl_scales.R
\name{pwfsl_scales}
\alias{pwfsl_scales}
\title{Add PWFSL PM25 timeseries scales}
\usage{
pwfsl_scales(data = NULL, startdate = NULL, enddate = NULL,
  ylim = NULL)
}
\arguments{
\item{data}{pm25 timeseries data. Should match the default dataset of the
plot}

\item{startdate}{Desired startdate for x-axis, in a format that can be 
parsed with \link{parseDatetime}.}

\item{enddate}{Desired enddate for x-axis, in a format that can be parsed 
with \link{parseDatetime}.}

\item{ylim}{custom y-axis limits. This function will apply a default limit
depending on the data.}
}
\description{
Add PWFSL-style x-axis and y-axis scales suitable for a timeseries plot 
showing PM2.5 data.
}
