% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scale_x_pwfsl_date.R
\name{scale_x_pwfslDate}
\alias{scale_x_pwfslDate}
\title{Add date scale with custom styling for timezone and daterange}
\usage{
scale_x_pwfslDate(startdate = NULL, enddate = NULL, timezone = NULL,
  expand = c(0, 0.05), breaks = NULL, minor_breaks = NULL,
  date_labels = "\%b \%d", ...)
}
\arguments{
\item{startdate}{Desired axis start date limit, in a format that can be 
parsed with \link{parseDatetime}.}

\item{enddate}{Desired axis end date limit, in a format that can be 
parsed with \link{parseDatetime}.}

\item{timezone}{Timezone for label formatting.}

\item{expand}{Vector of range expansion constants used to add some padding 
around the data, to ensure that they are placed some distance away from the 
axes.}

\item{breaks}{Custom breaks. If NULL, suitable breaks are calculated.}

\item{minor_breaks}{Custom minor breaks. If NULL, suitable breaks are
calculated.}

\item{date_labels}{date format string for formatting date labels.}
}
\description{
Add a date scale and custom formatting for creating 
consistent timeseries plots.
}
