% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_nowcast.R
\name{stat_nowcast}
\alias{stat_nowcast}
\title{Add NowCast values to a plot}
\usage{
stat_nowcast(
  mapping = NULL,
  data = NULL,
  version = "pm",
  includeShortTerm = FALSE,
  geom = "path",
  aqiColors = FALSE,
  mv4Colors = FALSE,
  position = "identity",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If
specified and \code{inherit.aes = TRUE} (the default), it is combined with
the default mapping at the top level of the plot. You must supply
\code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options:
if \code{NULL}, the default, the data is inherited from the plot data. A
\code{data.frame} or other object, will override the plot data. A
\code{function} will be called with a  single argument, the plot data. The
return value must be a \code{data.frame}, and will be used as the layer
data.}

\item{version}{character identity specifying the type of nowcast algorithm to
be used. For details see \link[AirMonitor]{monitor_nowcast}.}

\item{includeShortTerm}{calculate preliminary NowCast values starting with
the 2nd hour.}

\item{geom}{The geometic object to display the data}

\item{aqiColors}{if \code{TRUE}, AQI colors will be displayed.}

\item{mv4Colors}{if \code{TRUE}, AQI colors from the monitoring V4 site will
be used. Ignored if \code{aqiColors} is \code{FALSE}.}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function.}

\item{na.rm}{remove NA values from data}

\item{show.legend}{logical indicating whether this layer should be included
in legends.}

\item{inherit.aes}{if \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and the aesthetics and shouldn't inherit behaviour from
the default plot specificatino, eg \code{borders()}.}

\item{...}{additional arguments passed on to \code{layer()}, such as
aesthetics.}
}
\description{
This function calculates the NowCast version of the data, and adds it to a
plot. The default is to add a NowCast line.
}
\examples{
library(AirMonitorPlots)

ggplot_pm25Timeseries(
  AirMonitor::Carmel_Valley,
  startdate = 20160801,
  enddate = 20160810
) +
  geom_pm25Points() +
  stat_nowcast()
}
