% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_airfire}
\alias{theme_airfire}
\title{Theme for AirFire plots}
\usage{
theme_airfire(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A \emph{ggplot} theme.
}
\description{
Applies the package standard theme to a \emph{ggplot} plot object.
}
