% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_custom_size}
\alias{theme_custom_size}
\title{Custom styling for plots based on size}
\usage{
theme_custom_size(size = "large")
}
\arguments{
\item{size}{\code{small} or \code{large}. \code{style = small} is appropriate
for plots 450x450px or smaller; \code{style = large} is appropriate for
plots larger than 450x450px.}
}
\value{
A \emph{ggplot} theme.
}
\description{
Applies a theme to a \emph{ggplot} plot object. This theme is intended for
use when generating plots for the AirFire monitoring site.
}
\keyword{internal}
