% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_dailyBarplot_airfire}
\alias{theme_dailyBarplot_airfire}
\title{Theme for AirFire daily barplot for use in the monitoring site}
\usage{
theme_dailyBarplot_airfire(size = "large")
}
\arguments{
\item{size}{\code{small} or \code{large}. \code{style = small} is appropriate
for plots 450x450px or smaller; \code{style = large} is appropriate for
plots larger than 450x450px.}
}
\value{
A \emph{ggplot} theme.
}
\description{
Applies a theme to a \emph{ggplot} plot object. This theme is intended for
use with the \code{\link{monitor_ggDailyBarplot}} function and generates
plots suitable for the AirFire monitoring site. It is suited to display of 1-4
weeks of data.
}
