% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dailyBarplot_monitoring.R
\name{theme_dailyBarplot_monitoring}
\alias{theme_dailyBarplot_monitoring}
\title{Theme for "monitoring site" daily barplots}
\usage{
theme_dailyBarplot_monitoring(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A **ggplot** theme
}
\description{
Applies a theme to a **ggplot** plot object.
This theme is intended for use with the `dailyBarplot()` function and
generates plots suitable for the PWFSL monitoring site. It is suited to
display of 1-4 weeks of data.
}
