% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_dailyByHour_airfire}
\alias{theme_dailyByHour_airfire}
\title{Theme for AirFire dailyByHour plot for use in the monitoring site}
\usage{
theme_dailyByHour_airfire(size = "large")
}
\arguments{
\item{size}{\code{small} or \code{large}. \code{style = small} is appropriate
for plots 450x450px or smaller; \code{style = large} is appropriate for
plots larger than 450x450px.}
}
\value{
A \emph{ggplot} theme
}
\description{
Applies a theme to a \emph{ggplot} plot object. This theme is intended for
use with the \code{\link{monitor_ggDailyByHour}} function and generates plots
suitable for the AirFire monitoring site.
}
