% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_dailyHourlyBarplot_airfire}
\alias{theme_dailyHourlyBarplot_airfire}
\title{Theme for AirFire Daily Hourly Barplot}
\usage{
theme_dailyHourlyBarplot_airfire(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A **ggplot** theme
}
\description{
Applies a theme to a **ggplot** plot object. This is a minimal theme tailored
for faceted bar plots.
}
