% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mazamaBar.R
\name{theme_mazamaBar}
\alias{theme_mazamaBar}
\title{Theme tailored for use with bar plots in the style of Mazama Science}
\usage{
theme_mazamaBar(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A **ggplot** theme
}
\description{
Applies a theme to a **ggplot** plot object. This is a minimal theme tailored
for faceted bar plots.
}
