% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_theme_timeseriesPlot_pwfsl.R,
%   R/theme_timeseriesPlot_pwfsl.R
\name{theme_timeseriesPlot_pwfsl}
\alias{theme_timeseriesPlot_pwfsl}
\title{Theme for "monitoring site" timeseries barplots}
\usage{
theme_timeseriesPlot_pwfsl(base_size = 11, base_family = "")

theme_timeseriesPlot_pwfsl(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}

\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A **ggplot** theme

A **ggplot** theme
}
\description{
Applies a theme to a **ggplot** plot object.
This theme is intended for use with the `timeseriesPlot()` function and
generates plots suitable for the PWFSL monitoring site. It is suited to
display of 1-4 weeks of data.

Applies a theme to a **ggplot** plot object.
This theme is intended for use with the `timeseriesPlot()` function and
generates plots suitable for the PWFSL monitoring site. It is suited to
display of 1-4 weeks of data.
}
