% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ggClockPlot.R
\name{tidy_ggClockPlot}
\alias{tidy_ggClockPlot}
\title{Create a "clock" plot for a single monitor}
\usage{
tidy_ggClockPlot(ws_tidy, startdate = NULL, enddate = NULL,
  monitorID = NULL, timezone = NULL, ...)
}
\arguments{
\item{ws_tidy}{dataframe of monitor data, created from a \code{ws_monitor}
object using \code{monitor_toTidy()}.}

\item{startdate}{Desired startdate for data to include, in a format that can be 
parsed with \link{parseDatetime}.}

\item{enddate}{Desired enddate for data to include, in a format that can be parsed 
with \link{parseDatetime}.}

\item{monitorID}{monitorID to include in the plot.}

\item{timezone}{Timezone for x-axis scale. If NULL and only one timezone present
in the data, the data timezone will be used. If NULL and multiple timezones 
present, the default is UTC.}

\item{...}{Additional arguments passed on to \code{\link{custom_pm25DiurnalScales}}.}
}
\value{
A **ggplot** object
}
\description{
This function assembles various layers to create a production-ready
"clock" plot for one monitor.
}
\examples{
\dontrun{
ws_monitor <- airnow_loadLatest()
ws_tidy <- monitor_toTidy(ws_monitor)
tidy_ggClockPlot(ws_tidy, monitorID = "060631010_01")
}
}
