% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ggDailyBarplot.R
\name{tidy_ggDailyBarplot}
\alias{tidy_ggDailyBarplot}
\title{Create a daily barplot timeseries for one or more monitors}
\usage{
tidy_ggDailyBarplot(ws_tidy, startdate = NULL, enddate = NULL,
  monitorIDs = NULL, title = NULL, timezone = NULL, today = TRUE)
}
\arguments{
\item{ws_tidy}{dataframe of monitor data, created from a \code{ws_monitor}
object using \code{monitor_toTidy()}.}

\item{startdate}{Desired start date (integer or character in ymd format or POSIXct)}

\item{enddate}{Desired end date (integer or character in ymd format or POSIXct)}

\item{monitorIDs}{vector of monitorIDs to include in the plot. If 
more than one, different monitors will be plotted in different colors.}

\item{title}{Plot title. If NULL, a suitable title will be constructed.}

\item{timezone}{Timezone for x-axis scale. If NULL and only one timezone present
in the data, the data timezone will be used. If NULL and multiple timezones 
present, the default is UTC.}
}
\value{
A **ggplot** object
}
\description{
This function assembles various layers to create a production-ready
daily barplot for one or more monitors.
}
\examples{
ws_monitor <- PWFSLSmoke::Carmel_Valley
ws_tidy <- monitor_toTidy(ws_monitor)
tidy_ggDailyBarplot(ws_tidy)
}
