% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ggTimeseries.R
\name{tidy_ggTimeseries}
\alias{tidy_ggTimeseries}
\title{Create timeseries plot for one or more monitors}
\usage{
tidy_ggTimeseries(ws_tidy, startdate = NULL, enddate = NULL,
  style = NULL, aqiStyle = NULL, monitorIDs = NULL, title = NULL)
}
\arguments{
\item{ws_tidy}{dataframe of monitor data, created from a \code{ws_monitor}
object using \code{monitor_toTidy()}.}

\item{startdate}{Desired start date (integer or character in ymd format or POSIXct)}

\item{enddate}{Desired end date (integer or character in ymd format or POSIXct)}

\item{style}{Plot style. Not currently supported.}

\item{aqiStyle}{AQI style to add AQI color bars, lines, and labels. 
Not currently supported.}

\item{monitorIDs}{vector of monitorIDs to include in the plot. If 
more than one, different monitors will be plotted in different colors.}

\item{title}{Plot title. If NULL, a suitable title will be constructed.}
}
\value{
A **ggplot** object
}
\description{
This function assembles various layers to create a production-ready
timeseries plot for one or more monitors.
}
