% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesPlot.R
\name{timeseriesPlot}
\alias{timeseriesPlot}
\title{Create a timeseries plot for a single monitor}
\usage{
timeseriesPlot(ws_monitor, startdate = NULL, enddate = NULL,
  style = NULL, aqiStyle = NULL, monitorID = NULL, title = "")
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object.}

\item{startdate}{Desired start date (integer or character in Ymd format
or \code{POSIXct}).}

\item{enddate}{Desired end date (integer or character in Ymd format
or \code{POSIXct}).}

\item{style}{Plot style.}

\item{aqiStyle}{AQI style to add AQI color bars, lines and labels.}

\item{monitorID}{Monitor ID of interest. Required if \code{ws_monitor} 
contains more than one monitor.}

\item{title}{Optional title for the plot.}
}
\value{
A `ggplot` plot object with a "daily barplot" for a single monitor.
}
\description{
Create a  timeseries plot of PM2.5 values at a location.

This function presents a simplified interface to 
\code{\link{timeseriesPlotBase}} and collects common options into a set of 
named styles.

Current styles include:
\itemize{
\item{\code{pwfsl} -- used in the PWFSL monitoring site. Best for 1-2
weeks of data.}
}

The returned object may be further amended with **ggplot** elements.
}
\details{
Daily averages are calculated from midnight-to-midnight in monitor 
local time.
}
\examples{
NW <- PWFSLSmoke::Northwest_Megafires
timeseriesPlot(NW, "2015-08-01", "2015-08-14",
               style = "pwfsl",
               monitorID = "160690014_01",
               title = "Daily Average PM2.5\\nRuebens, Idaho")
}
\seealso{
\code{\link{timeseriesPlotBase}}
}
