% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesPlotBase.R
\name{timeseriesPlotBase}
\alias{timeseriesPlotBase}
\title{Create a timeseries plot for one or more monitors}
\usage{
timeseriesPlotBase(ws_monitor, startdate = NULL, enddate = NULL,
  colorPalette = aqiPalette("aqi"), dataStyle = "pwfsl",
  nowcastStyle = "pwfsl", ylimStyle = "auto", aqiStyle = NULL,
  dateFormat = "\%b \%d", title = "")
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object.}

\item{startdate}{Desired start date (integer or character in Ymd format 
or \code{POSIXct}).}

\item{enddate}{Desired end date (integer or character in Ymd format
or \code{POSIXct}).}

\item{colorPalette}{Palette function to convert monitor values into colors.}

\item{dataStyle}{Style for displaly of PM2.5 values. One of \code{pwfsl}.}

\item{nowcastStyle}{Style for displaly of Nowcast values. One of \code{pwfsl}.}

\item{ylimStyle}{Style of y-axis limits. One of \code{auto|pwfsl}.}

\item{aqiStyle}{AQI style to add AQI color bars, lines and labels.}

\item{dateFormat}{Format for x-axis dates. Used for \code{date_labels}
argument to \code{scale_x_datetime}.}

\item{title}{Optional title.}
}
\value{
A `ggplot` plot object with a daily bar plot for a single monitor.
}
\description{
Create a timeseries plot with points or lines showing PM2.5 data over a 
period for one or more monitors in a \emph{ws_monitor} object.
}
