% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windBarbs.R
\name{wind_barb}
\alias{wind_barb}
\title{Add wind barbs to a map}
\usage{
wind_barb(x, y, speed, dir, circleSize = 1, circleFill = "transparent",
  lineCol = 1, barbSize = 1, ...)
}
\arguments{
\item{x}{vector of longitudes}

\item{y}{vector of latitudes}

\item{speed}{wind speed in knots (Doesn't currently support vectorized wind speed)}

\item{dir}{wind direction in degrees clockwise from north}

\item{circleSize}{size of the circle}

\item{circleFill}{circle fill color (currently not supported)}

\item{lineCol}{line color (currently not supported)}

\item{barbSize}{size of the barb}

\item{...}{additional arguments to be passed to \code{lines}}
}
\description{
Add a multi-sided polygon to a plot.
}
\examples{
maps::map('state', "washington")
x <- c(-121, -122)
y <- c(47.676057, 47)
wind_barb(x, y, speed = 125, dir = c(45, 90), lwd = 2, circleSize = .5, barbSize = 1.5)
}
\keyword{plotting}
