% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brandPlot.R
\name{brandPlot}
\alias{brandPlot}
\title{Add a logo to a ggplot object}
\usage{
brandPlot(
  plot,
  brandStyle = "logo",
  brandName = "MazamaScience",
  brandFilePath = NULL,
  location = c("topright", "topleft", "bottomright", "bottomleft")[1],
  size = 0.1
)
}
\arguments{
\item{plot}{\code{ggplot} object}

\item{brandStyle}{\code{"logo"} or \code{"icon"}.}

\item{brandName}{Name of brand: \code{"MazamaScience"}, \code{"USFS"}, or
\code{"AirFire"}.}

\item{brandFilePath}{Path to brand logo or icon. If not \code{NULL},
overrides \code{brandStyle} and \code{brandName}.}

\item{location}{String indicating the location where the logo should be
printed. Options are: \code{"topright"}, \code{"topleft"},
\code{"bottomright"}, or \code{"bottomleft"}.}

\item{size}{Brand icon or logo width, in fraction of plot width.}
}
\value{
A \code{gTree} object, which can be printed with \code{grid.draw()}.
}
\description{
Add a logo to a ggplot object.
}
\examples{
mts_monitor <- AirMonitor::Carmel_Valley
mts_tidy <- monitor_toTidy(mts_monitor)
p <- ggplot_pm25Timeseries(mts_tidy) +
  stat_dailyAQCategory(adjustylim = TRUE)
brandPlot(p, location = "topright", size = .2)
brandPlot(p, location = "bottomright", brandName = "USFS")
brandPlot(p, brandName = "AirFire", location = "topleft", size = .15)
}
