% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brandPlot.R
\name{brandPlot}
\alias{brandPlot}
\title{Add Branding to a ggplot object}
\usage{
brandPlot(plot, fullLogo = TRUE, location = c("topright", "topleft",
  "bottomright", "bottomleft")[1])
}
\arguments{
\item{plot}{\code{ggplot} object}

\item{fullLogo}{if \code{TRUE}, the full Mazama Science logo is added. If 
\code{FALSE}, only the icon will be added.}

\item{location}{string indicating the location where the logo should be 
printed. Options are: \code{"topright"}, \code{"topleft"}, \code{"bottomright"}, 
or \code{"bottomleft"}.}
}
\value{
A \code{gTree} object, which can be printed with \code{grid.draw()}.
}
\description{
Add a logo to a ggplot object.
}
\examples{
ws_monitor <- PWFSLSmoke::Carmel_Valley
ws_tidy <- monitor_toTidy(ws_monitor)
p <- ggplot_pm25Timeseries(ws_tidy) +
  stat_dailyAQILevel(adjustylim = TRUE) 
brandPlot(p, location = "bottomright")
}
