% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockPlot.R
\name{clockPlot}
\alias{clockPlot}
\title{Create a Clock plot for a single monitor}
\usage{
clockPlot(ws_monitor, monitorID = NULL, startdate = NULL,
  style = "icon", centerColor = "transparent")
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object.}

\item{monitorID}{Monitor ID of interest.}

\item{startdate}{Desired start date (integer or character in Ymd format).}

\item{style}{Plot style one of \code{base|icon|icon_fan|full|full_fan}.}

\item{centerColor}{Color used for the center of the circle.}
}
\value{
A **ggplot** plot object with a "clock plot" for a single monitor.
}
\description{
Create a "clock plot" showing PM2.5 data for a single day for the given 
monitors. A colored bar curves around in a clockwise manner with 12/4 of the
bar colored for each hour of the local time day.
}
\examples{
ws_monitor <- PWFSLSmoke::Carmel_Valley
monitorID <- ws_monitor$meta$monitorID[1]
startdate <- "2016-08-07"
clockPlot(ws_monitor, monitorID, startdate, style = "icon")
}
