% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarnayPlot.R
\name{createTarnayPlot}
\alias{createTarnayPlot}
\title{Create a Tarnay plot for many monitors}
\usage{
createTarnayPlot(monitors, data, columns = 1, title = NULL, xLabel = NULL,
  yLabel = NULL, includeLink = TRUE, hourlyType = "nowcast",
  colorScale = "epa_aqi", includeThirdCol = FALSE)
}
\arguments{
\item{monitors}{Monitor ID(s) to create plot for.}

\item{data}{Data used to create plot (NOTE: currently must be a `ws_monitor`
object).}

\item{columns}{Number of columns the faceted plot should have (default 1).}

\item{title}{The title of the plot. Defaults to specifying the types of
data present in the plot.}

\item{xLabel}{The x-axis label of the plot. Defaults to years present in
data.}

\item{yLabel}{The y-axis label of the plot. Defaults to PM2.5.}

\item{includeLink}{Option to include a link to an AQI explainer at the bottom
of the plot (default `TRUE`).}

\item{hourlyType}{The type of hourly data to include in the plot. The options
include "nowcast" (hourly nowcast values), "raw" (raw hourly values), or
"none" (no hourly data at all) (default "nowcast").}

\item{colorScale}{The ordered color pallete used to represent each AQI
category. Currently defaults to (and only accepts) "epa_aqi".}

\item{includeThirdCol}{Option to include a third column in the legend.
Currently in testing (default `False`).}
}
\value{
A **ggplot** plot of the given monitors and data.
}
\description{
Create a timeseries barplot showing PM2.5 data for the given monitors. The
overall plot is faceted by monitor, and each facet has two sets of columns:
one for daily levels, and one for hourly levels.
}
\examples{

}
