% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarnayPlot.R
\name{createTarnayPlot}
\alias{createTarnayPlot}
\title{Create a summary barplot with multiple bar types}
\usage{
createTarnayPlot(monitors, data, columns = 1)
}
\arguments{
\item{monitors}{Monitor ID(s) to create plot for.}

\item{data}{Data used to create plot.}

\item{columns}{Number of columns the faceted plot should have (default 1).}
}
\value{
A `ggplot` barplot of the given monitors and data.
}
\description{
Creates a faceted barplot with multiple column types and
    one or more monitor IDs.
}
\examples{

}
