% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_pm25DiurnalScales.R
\name{custom_pm25DiurnalScales}
\alias{custom_pm25DiurnalScales}
\title{PWFSL PM2.5 timeseries scales}
\usage{
custom_pm25DiurnalScales(data = NULL, ylim = NULL, xlab = NULL,
  ylab = "PM2.5 (µg/m3)", yexp = c(0.05, 0), xexp = c(0, 0.05))
}
\arguments{
\item{data}{pm25 timeseries data. Should match the default dataset of the
plot}

\item{ylim}{custom y-axis limits. This function will apply a default limit
depending on the data.}

\item{xlab}{Custom x-axis label. If \code{NULL} a default xlab will be generated.}

\item{ylab}{Custam y-axis label.}

\item{yexp}{Vector of range expansion constants used to add some padding around the 
data on the y-axis, to ensure that they are placed some distance away from the axes.}

\item{xexp}{Vector of range expansion constants used to add some padding around the 
data on the x-axis, to ensure that they are placed some distance away from the axes.}
}
\description{
Add PWFSL-style x-axis and y-axis scales suitable for a timeseries plot 
showing PM2.5 data.
}
