% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyBarplot.R
\name{dailyBarplot}
\alias{dailyBarplot}
\title{Create a daily barplot for a single monitor}
\usage{
dailyBarplot(ws_monitor, monitorID = NULL, startdate = NULL,
  enddate = NULL, style = "monitoring", currentNowcast = NULL,
  currentPrediction = NULL, title = "")
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object.}

\item{monitorID}{Monitor ID of interest.}

\item{startdate}{Desired start date (integer or character in Ymd format
or \code{POSIXct}).}

\item{enddate}{Desired end date (integer or character in Ymd format
or \code{POSIXct}).}

\item{style}{Plot style.}

\item{currentNowcast}{Real-time current Nowcast value -- for use in plots 
presented in the PWFSL monitoring site.}

\item{currentPrediction}{Real-time current prediction for today's daily 
average -- for use in plots presented in the PWFSL monitoring site.}

\item{title}{Optional title for the plot.}
}
\value{
A **ggplot** plot object with a "daily barplot" for a single monitor.
}
\description{
Create a daily average barplot of PM2.5 values at a location.

This function presents a simplified interface to \code{\link{dailyBarplotBase}}
and collects common options into a set of named styles.

Current styles include:
\itemize{
\item{\code{monitoring} -- used in the PWFSL monitoring site. Best for 1-2
weeks of data.}
}

The returned object may be further amended with **ggplot** elements.
}
\details{
Daily averages are calculated from midnight-to-midnight in monitor 
local time.
}
\examples{
mon <- PWFSLSmoke::Carmel_Valley
id <- mon$meta$monitorID[1]
start1 <- "2016-06-01"
start2 <- "2016-07-01"
start3 <- "2016-08-01"
end1 <- "2016-08-08"
end2 <- "2016-08-15"
end3 <- "2016-08-25"
dbp1 <- dailyBarplot(mon, id, start3, end1, "monitoring",
                     title = "Carmel Valley\\n2016")
dbp2 <- dailyBarplot(mon, id, start3, end2, "monitoring",
                     title = "Carmel Valley\\n2016")
dbp3 <- dailyBarplot(mon, id, start2, end2, "monitoring",
                     title = "Carmel Valley\\n2016")
dbp4 <- dailyBarplot(mon, id, start1, end3, "monitoring",
                     title = "Carmel Valley\\n2016")
gridExtra::grid.arrange(dbp1, dbp2, dbp3, dbp4, nrow = 2)
}
\seealso{
\code{\link{dailyBarplotBase}}
}
