% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyBarplot.R
\name{dailyBarplot}
\alias{dailyBarplot}
\title{Create a daily barplot for a single monitor}
\usage{
dailyBarplot(ws_monitor, startdate = NULL, enddate = NULL,
  style = "pwfsl", monitorID = NULL, currentNowcast = NULL,
  currentPrediction = NULL, title = "")
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object.}

\item{startdate}{Desired start date (integer or character in Ymd format
or \code{POSIXct}).}

\item{enddate}{Desired end date (integer or character in Ymd format
or \code{POSIXct}).}

\item{style}{Plot style.}

\item{monitorID}{Monitor ID of interest. Required if \code{ws_monitor} 
contains more than one monitor.}

\item{currentNowcast}{Real-time current Nowcast value -- for use in plots 
presented in the PWFSL monitoring site.}

\item{currentPrediction}{Real-time current prediction for today's daily 
average -- for use in plots presented in the PWFSL monitoring site.}

\item{title}{Optional title for the plot.}
}
\value{
A **ggplot** plot object with a "daily barplot" for a single monitor.
}
\description{
Create a daily average barplot of PM2.5 values at a location.

This function presents a simplified interface to \code{\link{dailyBarplotBase}}
and collects common options into a set of named styles.

Current styles include:
\itemize{
\item{\code{pwfsl} -- used in the PWFSL monitoring site. Best for 1-2
weeks of data.}
}

The returned object may be further amended with **ggplot** elements.
}
\details{
Daily averages are calculated from midnight-to-midnight in monitor 
local time.
}
\examples{
NW <- PWFSLSmoke::Northwest_Megafires
start1 <- "2015-08-20"
end1 <- "2015-08-27"
start2 <- "2015-06-01"
end2 <- "2015-11-01"
dbp1 <- dailyBarplot(NW, start1, end1, "pwfsl",
                     monitorID = "160690014_01",
                     title = "Ruebens, Idaho\\n201")
dbp2 <- dailyBarplot(NW, start2, end2, "month",
                     monitorID = "160690014_01",
                     title = "Ruebens, Idaho\\n2015")
gridExtra::grid.arrange(dbp1, dbp2, nrow = 1)
}
\seealso{
\code{\link{dailyBarplotBase}}
}
