% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyBarplotBase.R
\name{dailyBarplotBase}
\alias{dailyBarplotBase}
\title{Create a daily bar plot for a single monitor}
\usage{
dailyBarplotBase(ws_monitor, startdate = NULL, enddate = NULL,
  colorPalette = aqiPalette("aqi"), ylimStyle = "auto",
  borderColor = "black", borderSize = 1, currentNowcast = NULL,
  currentPrediction = NULL, showAQIStackedBars = FALSE,
  showAQILines = FALSE, showAQILegend = FALSE, title = "")
}
\arguments{
\item{ws_monitor}{\emph{ws_monitor} object containing a single monitor.}

\item{startdate}{Desired start date (integer or character in Ymd format 
or \code{POSIXct}).}

\item{enddate}{Desired end date (integer or character in Ymd format
or \code{POSIXct}).}

\item{colorPalette}{Palette function to convert monitor values into colors.}

\item{ylimStyle}{Style of y-axis limits. One of \code{auto|pwfsl}.}

\item{borderColor}{Border color for individual bars.}

\item{borderSize}{Border size for individual bars.}

\item{currentNowcast}{Real-time current Nowcast value -- for use in plots 
presented in the PWFSL monitoring site.}

\item{currentPrediction}{Real-time current prediction for today's daily 
average -- for use in plots presented in the PWFSL monitoring site.}

\item{title}{Optional title.}
}
\value{
A **ggplot** plot object with a daily bar plot for a single monitor.
}
\description{
Create a bar plot showing daily average PM2.5 data over a period for the given 
monitor. Colored bars represent the PM2.5 readings for each day.
}
\examples{
ws_monitor <- PWFSLSmoke::Carmel_Valley
startdate <- "2016-08-5"
enddate <- "2016-08-19"
dailyBarplotBase(ws_monitor, startdate, enddate)
}
