% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_geom_pm25Points.R
\name{geom_pm25Points}
\alias{geom_pm25Points}
\title{Add pm25 points to a plot}
\usage{
geom_pm25Points(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE,
  stat = "identity", timeseries.legend = FALSE,
  legend.label = "PM2.5", ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{timeseries.legend}{Logical indicating whether to set mappings so that
a legend can later be added using \link{legend_pm25Timeseries}.}

\item{legend.label}{Label for pm25Points part of the legend. Must match 
legend.labels argument in \link{legend_pm25Timeseries}.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
Add pm25 points to a ggplot. This is a wrapper around \code{geom_point} with
default aesthetics. The main difference is the timeseries.legend option which,
when true, will set mappings so that a legend can be drawn for these points using
\link{legend_pm25Timeseries}.
}
