% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_pm25Diurnal.R
\name{ggplot_pm25Diurnal}
\alias{ggplot_pm25Diurnal}
\title{Create a pm25 diurnal ggplot}
\usage{
ggplot_pm25Diurnal(
  monitor,
  startdate = NULL,
  enddate = NULL,
  timezone = NULL,
  ylim = NULL,
  shadedNight = TRUE,
  mapping = aes_(x = ~hour, y = ~pm25),
  base_size = 11,
  ...
)
}
\arguments{
\item{monitor}{Default dataset to use when adding layers. Must be either a
\emph{mts_monitor} object or \code{mts_tidy} object.}

\item{startdate}{Desired startdate for data to include, in a format that can
be parsed with \link{parseDatetime}.}

\item{enddate}{Desired enddate for data to include, in a format that can be
parsed with \link{parseDatetime}.}

\item{timezone}{Timezone to use to set hours of the day.}

\item{ylim}{custom y-axis limits. This function will apply a default limit
depending on the data.}

\item{shadedNight}{add nighttime shading based on of middle day in selected
period.}

\item{mapping}{Default mapping for the plot.}

\item{base_size}{Base font size for theme.}

\item{...}{Additional arguments passed on to
\code{\link{custom_pm25DiurnalScales}}.}
}
\description{
Create a plot with default mappings and styling. Layers can then
be added to this plot using \code{ggplot2} syntax.
}
\examples{
monitor <- AirMonitor::Carmel_Valley
ggplot_pm25Diurnal(monitor) +
  coord_polar() +
  geom_pm25Points() +
  custom_aqiStackedBar(width = 1, alpha = .3)

ggplot_pm25Diurnal(
  monitor,
  startdate = 20160801,
  enddate = 20160810
) +
  stat_boxplot(aes(group = hour))

}
