% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_pm25Diurnal.R
\name{ggplot_pm25Diurnal}
\alias{ggplot_pm25Diurnal}
\title{Instantiate a pm25 timeseries ggplot}
\usage{
ggplot_pm25Diurnal(ws_data, startdate = NULL, enddate = NULL,
  timezone = NULL, ylim = NULL)
}
\arguments{
\item{ws_data}{Default dataset to use when adding layers. Must be either a \code{ws_monitor} 
object or \code{ws_tidy} object.}

\item{startdate}{Desired startdate for x-axis, in a format that can be 
parsed with \link{parseDatetime}.}

\item{enddate}{Desired enddate for x-axis, in a format that can be parsed 
with \link{parseDatetime}.}

\item{timezone}{Timezone for x-axis scale. If NULL and only one timezone present
in the data, the data timezone will be used. If NULL and multiple timezones 
present, the default is UTC.}

\item{ylim}{custom y-axis limits. This function will apply a default limit
depending on the data.}
}
\description{
Create a plot using ggplot with default mappings and styling. Layers can then 
be added to this plot using \code{ggplot2} syntax.
}
\examples{
ws_monitor <- PWFSLSmoke::Carmel_Valley
ggplot_pm25Diurnal(ws_monitor) +
  coord_polar() +
  geom_pm25Points() +
  custom_aqiStackedBar(width = 1, alpha = .3) 

ggplot_pm25Diurnal(ws_monitor, 
                   startdate = 20160801, 
                   enddate = 20160810) + 
stat_boxplot(aes(group = hour)) 
}
