% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_pm25Timeseries.R
\name{ggplot_pm25Timeseries}
\alias{ggplot_pm25Timeseries}
\title{Instantiate a pm25 timeseries ggplot}
\usage{
ggplot_pm25Timeseries(ws_data)
}
\arguments{
\item{ws_data}{Default dataset to use when adding layers. Must be either a \code{ws_monitor} 
object or \code{ws_tidy} object.}
}
\description{
Create a plot using ggplot with default mappings and styling. Layers can then 
be added to this plot using \code{ggplot2} syntax.
}
