% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_pm25Timeseries.R
\name{ggplot_pm25Timeseries}
\alias{ggplot_pm25Timeseries}
\title{Instantiate a pm25 timeseries ggplot}
\usage{
ggplot_pm25Timeseries(
  monitor,
  startdate = NULL,
  enddate = NULL,
  timezone = NULL,
  ylim = NULL,
  base_size = 11,
  ...
)
}
\arguments{
\item{monitor}{Monitoring data object to use when adding layers. Must be of
class \code{mts_monitor} or \code{mts_tidy}.}

\item{startdate}{Desired start date (integer or character in ymd format or
POSIXct).}

\item{enddate}{Desired end date (integer or character in ymd format or
POSIXct).}

\item{timezone}{Timezone for x-axis scale. If NULL and only one timezone
present in \code{monitor}, that timezone will be used. If NULL and multiple
timezones present, the default is UTC.}

\item{ylim}{custom y-axis limits. This function will apply a default limit
depending on the data.}

\item{base_size}{Base font size for theme.}

\item{...}{Additional arguments passed on to
\code{\link{custom_datetimeScale}}.}
}
\description{
Create a plot using ggplot with default mappings and styling. Layers can then
be added to this plot using \code{ggplot2} syntax.
}
\examples{
library(AirMonitorPlots)

AirMonitor::Carmel_Valley \%>\%
  ggplot_pm25Timeseries() +
  geom_point(shape = "square", alpha = .4)

}
