% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{isTidy}
\alias{isTidy}
\title{Check if Data has a 'Tidy' Format}
\usage{
isTidy(data = NULL)
}
\arguments{
\item{data}{Data to interrogate}
}
\value{
True if the data is in a recognized 'Tidy' format, otherwise False
}
