% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_ggCalendar.R
\name{monitor_ggCalendar}
\alias{monitor_ggCalendar}
\title{Calendar plot}
\usage{
monitor_ggCalendar(
  ws_monitor = NULL,
  monitorID = NULL,
  ncol = 4,
  title = NULL,
  discrete = TRUE,
  breaks = NULL,
  break_labels = NULL,
  aspect_ratio = 1,
  legend_title = NULL,
  stat = "mean"
)
}
\arguments{
\item{ws_monitor}{a \emph{ws_monitor} object.}

\item{monitorID}{a monitorID in the \emph{ws_monitor} object.}

\item{ncol}{the amount of columns in the plot.}

\item{title}{an optional title.}

\item{discrete}{a boolean that determines the color scale.}

\item{breaks}{the color scale break points.}

\item{break_labels}{the scale breaks names.}

\item{aspect_ratio}{the plot aspect ratio.}

\item{legend_title}{an optional title used in the legend.}

\item{stat}{the statistic used for daily aggregation (default: "mean").}
}
\value{
ggobject
}
\description{
Annual calendar view of a daily reading using a selected input.
}
