% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_ggDailyBarplot.R
\name{monitor_ggDailyBarplot}
\alias{monitor_ggDailyBarplot}
\title{Create a daily barplot for one or more monitors}
\usage{
monitor_ggDailyBarplot(
  monitor,
  startdate = NULL,
  enddate = NULL,
  id = NULL,
  style = "small",
  title = NULL,
  timezone = NULL,
  today = TRUE,
  ...
)
}
\arguments{
\item{monitor}{A \emph{mts_monitor} object.}

\item{startdate}{Desired start date (integer or character in ymd format or
POSIXct).}

\item{enddate}{Desired end date (integer or character in ymd format or
POSIXct).}

\item{id}{deviceDeploymentID to include in the plot. This can be NULL if
\code{monitor} only has one unique deviceDeploymentID.}

\item{style}{String indicating plotting style. Either \code{"large"} or
\code{"small"}. \code{style = "large"} is suitable for plots larger than
450x450px, and \code{"small"} is suitable for plots 450x450px or smaller.}

\item{title}{Plot title. If NULL, a suitable title will be constructed.}

\item{timezone}{Olson timezone name for x-axis scale and date parsing. If
NULL the timezone of the specified monitor will be used.}

\item{today}{Logical indicating whether to include a shaded "current NowCast"
bar for Today. Ignored if data is not current.}

\item{...}{Arguments passed onto \code{\link{ggplot_pm25Timeseries}}.}
}
\value{
A \emph{ggplot} object
}
\description{
This function assembles various layers to create a production-ready
daily barplot for one or more monitors.

The full range of data in \code{monitor} will be used unless both
\code{startdate} and \code{enddate} are specified.
}
\examples{
library(AirMonitorPlots)

AirMonitor::Carmel_Valley \%>\%
  monitor_ggDailyBarplot(
    startdate = 20160801,
    enddate = 20160810
  )

}
