% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_ggDailyBarplot.R
\name{monitor_ggDailyBarplot}
\alias{monitor_ggDailyBarplot}
\title{Create timeseries plot for one or more monitors}
\usage{
monitor_ggDailyBarplot(ws_monitor, startdate = NULL, enddate = NULL,
  monitorIDs = NULL, style = "large", title = NULL,
  timezone = NULL, today = TRUE)
}
\arguments{
\item{ws_monitor}{A `ws_monitor` object}

\item{startdate}{Desired start date (integer or character in ymd format or POSIXct)}

\item{enddate}{Desired end date (integer or character in ymd format or POSIXct)}

\item{monitorIDs}{vector of monitorIDs to include in the plot. If 
more than one, different monitors will be plotted in different colors.}

\item{style}{String indicating plotting style. Either \code{"large"} or \code{"small"}.
\code{style = "large"} is suitable for plots larger than 450x450px, and \code{"small"}
is suitable for plots 450x450px or smaller.}

\item{title}{Plot title. If NULL, a suitable title will be constructed.}

\item{timezone}{Timezone for x-axis scale. If NULL and only one timezone present
in the data, the data timezone will be used. If NULL and multiple timezones 
present, the default is UTC.}

\item{today}{Logical indicating whether to include a shaded "current NowCast" bar 
for Today. Ignored if data is not current.}
}
\value{
A **ggplot** object
}
\description{
This function wraps \link{tidy_ggDailyBarplot}, accepting a `ws_monitor` 
object to assemble various layers to create a production-ready
timeseries plot for one or more monitors.
}
\examples{
ws_monitor <- airnow_loadLatest()
monitor_ggTimeseries(ws_monitor, monitorIDs = "410432002_01")
}
