% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_ggDailyHourlyBarplot.R
\name{monitor_ggDailyHourlyBarplot}
\alias{monitor_ggDailyHourlyBarplot}
\title{Create a Daily-Hourly plot for many monitors}
\usage{
monitor_ggDailyHourlyBarplot(
  monitor,
  startdate = NULL,
  enddate = NULL,
  id = NULL,
  columns = 1,
  title = NULL,
  timezone = NULL,
  xLabel = NULL,
  yLabel = NULL,
  hourlyDataType = c("nowcast", "raw", "none"),
  palette = "epa_aqi",
  includeLegend = TRUE
)
}
\arguments{
\item{monitor}{A \emph{mts_monitor} object.}

\item{startdate}{Desired start date (integer or character in ymd format or
POSIXct).}

\item{enddate}{Desired end date (integer or character in ymd format or
POSIXct).}

\item{id}{Optional vector of deviceDeploymentIDs used to filter the data.}

\item{columns}{Number of columns the faceted plot should have.}

\item{title}{The title of the plot. Defaults to specifying the types of
data present in the plot.}

\item{timezone}{Olson timezone name for x-axis scale and date parsing. If
NULL the timezone of the specified monitor will be used.}

\item{xLabel}{The x-axis label of the plot. Defaults to years present in
data.}

\item{yLabel}{The y-axis label of the plot. Defaults to PM2.5.}

\item{hourlyDataType}{The type of hourly data to include in the plot. The
options include "nowcast" (hourly nowcast values), "raw" (raw hourly values),
or "none" (no hourly data at all).}

\item{palette}{The ordered color palette used to represent each AQI
category. Currently defaults to (and only accepts) "epa_aqi".}

\item{includeLegend}{Option to include a legend..}
}
\value{
A \strong{ggplot} plot of the given monitors and data.
}
\description{
Create a time series barplot showing PM2.5 data for the given monitors. The
overall plot is faceted by monitor, and each facet has two sets of columns:
one for daily levels, and one for hourly levels.

The full range of data in \code{monitor} will be used unless both
\code{startdate} and \code{enddate} are specified.

The timezone specified or, if \code{timezone = NULL}, that of the first
monitor encountered will be used for all time axes.
}
\examples{
\dontrun{
library(AirMonitor)
library(AirMonitorPlots)

# Fail gracefully if any resources are not available
try({

SF_IDs <- c(
  "ccdef3f0f6591e77_060010009",
  "06c3f2a66f8b708e_060010012",
  "7157b3dbac7c2043_060010011",
  "060750fa7ae26987a72cc4_060750005005_01"
)
SF_daily <- monitor_loadDaily() \%>\% monitor_select(id = SF_IDs)
SF_latest <- monitor_loadLatest() \%>\% monitor_select(id = SF_IDs)
SF_full <- monitor_combine(SF_daily, SF_latest)
today <- lubridate::floor_date(lubridate::now('America/Los_Angeles'), unit='day')
now <- lubridate::floor_date(lubridate::now('America/Los_Angeles'), unit='hour')
starttime <- today - lubridate::ddays(4)
SF_4day <- monitor_filterDatetime(SF_full, starttime, now)

# Create plot using pre subset data
monitor_ggDailyHourlyBarplot(SF_4day, id = SF_IDs)

# Create plot using data subset by function
monitor_ggDailyHourlyBarplot(SF_full, starttime, now, SF_IDs)

}, silent = FALSE)
}
}
