% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_ggTimeseries.R
\name{monitor_ggTimeseries}
\alias{monitor_ggTimeseries}
\title{Create a timeseries plot for one or more monitors}
\usage{
monitor_ggTimeseries(ws_monitor, startdate = NULL, enddate = NULL,
  style = "small", aqiStyle = NULL, monitorIDs = NULL,
  title = NULL, ...)
}
\arguments{
\item{ws_monitor}{A \code{ws_monitor} object.}

\item{startdate}{Desired start date (integer or character in ymd format or
POSIXct).}

\item{enddate}{Desired end date (integer or character in ymd format or
POSIXct).}

\item{style}{Plot style. \code{small} or \code{large}. \code{style = small}
is appropriate for plots 450x450px or smaller; \code{style = large} is
appropriate for plots larger than 450x450px.}

\item{aqiStyle}{AQI style to add AQI color bars, lines, and labels. Not
currently supported.}

\item{monitorIDs}{vector of monitorIDs to include in the plot. If more than
one, different monitors will be plotted in different colors.}

\item{title}{Plot title. If NULL, a suitable title will be constructed.}

\item{...}{Arguments passed onto \code{\link{ggplot_pm25Timeseries}}.}
}
\value{
A \emph{ggplot} object.
}
\description{
This function assembles various layers to create a production-ready
timeseries plot for one or more monitors.
}
\examples{
NM <- PWFSLSmoke::Northwest_Megafires
monitor_ggTimeseries(
  NM,
  startdate = 20150815,
  enddate = 20150831,
  monitorIDs = "160690014_01"
)

CV <- PWFSLSmoke::Carmel_Valley
monitor_ggTimeseries(CV, startdate = 20160801, enddate = 20160810)

\dontrun{
ws_monitor <- airnow_loadLatest()
monitor_ggTimeseries(ws_monitor, monitorID = "410432002_01")
}
}
