% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_ggTimeseries.R
\name{monitor_ggTimeseries}
\alias{monitor_ggTimeseries}
\title{Create a timeseries plot for one or more monitors}
\usage{
monitor_ggTimeseries(
  monitor,
  startdate = NULL,
  enddate = NULL,
  id = NULL,
  style = c("small", "large"),
  title = NULL,
  timezone = NULL,
  ...
)
}
\arguments{
\item{monitor}{A \emph{mts_monitor} object.}

\item{startdate}{Desired start date (integer or character in ymd format or
POSIXct).}

\item{enddate}{Desired end date (integer or character in ymd format or
POSIXct).}

\item{id}{vector of deviceDeploymentIDs to include in the plot. If more than
one, different monitors will be plotted in different colors.}

\item{style}{Plot style. \code{small} or \code{large}. \code{style = small}
is appropriate for plots 450x450px or smaller; \code{style = large} is
appropriate for plots larger than 450x450px.}

\item{title}{Plot title. If NULL, a suitable title will be constructed.}

\item{timezone}{Olson timezone name for x-axis scale and date parsing. If
NULL the timezone of the specified monitor will be used.}

\item{...}{Arguments passed onto \code{\link{ggplot_pm25Timeseries}}.}
}
\value{
A \emph{ggplot} object.
}
\description{
This function assembles various layers to create a production-ready
timeseries plot for one or more monitors.
}
\examples{
library(AirMonitorPlots)

AirMonitor::NW_Megafires \%>\%
  monitor_ggTimeseries(
    startdate = 20150815,
    enddate = 20150831,
    id = "d01b991c53583048_160690012_03"
  )

AirMonitor::Carmel_Valley \%>\%
  monitor_ggTimeseries(
    startdate = 20160801,
    enddate = 20160810
  )

}
