% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_ggTimeseries.R
\name{monitor_ggTimeseries}
\alias{monitor_ggTimeseries}
\title{Create timeseries plot for one or more monitors}
\usage{
monitor_ggTimeseries(ws_monitor, startdate = NULL, enddate = NULL,
  style = NULL, aqiStyle = NULL, monitorIDs = NULL, title = NULL)
}
\arguments{
\item{ws_monitor}{A `ws_monitor` object}

\item{startdate}{Desired start date (integer or character in ymd format or POSIXct)}

\item{enddate}{Desired end date (integer or character in ymd format or POSIXct)}

\item{style}{Plot style. Not currently supported.}

\item{aqiStyle}{AQI style to add AQI color bars, lines, and labels. 
Not currently supported.}

\item{monitorIDs}{vector of monitorIDs to include in the plot. If 
more than one, different monitors will be plotted in different colors.}

\item{title}{Plot title. If NULL, a suitable title will be constructed.}
}
\value{
A **ggplot** object
}
\description{
This function wraps \link{tidy_ggTimeseries}, accepting a `ws_monitor` 
object to assemble various layers to create a production-ready
timeseries plot for one or more monitors.
}
