% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_isTidy.R
\name{monitor_isTidy}
\alias{monitor_isTidy}
\title{Check if data is tidy-formatted mts_monitor data}
\usage{
monitor_isTidy(data = NULL)
}
\arguments{
\item{data}{Data object to validate.}
}
\value{
TRUE if the data is in a recognized 'Tidy' format, otherwise FALSE
}
\description{
Verifies that the given data can be treated as tidy-formatted
data. This is done by verifying that the data is a tibble data.frame object
with columns for information in all \emph{mts_monitor} objects.
}
\examples{
library(AirMonitor)

mts_monitor <-
  NW_Megafires \%>\%
  AirMonitor::monitor_select(
    c('450d08fb5a3e4ea0_530470009', '40ffdacb421a5ee6_530470010')
  )

mts_tidy <- monitor_toTidy(mts_monitor)
monitor_isTidy(mts_tidy)

\dontrun{
monitor_isTidy(mts_monitor)
}

}
