% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_toTidy.R
\name{monitor_toTidy}
\alias{monitor_toTidy}
\title{Convert `ws_monitor` data to a tidy format}
\usage{
monitor_toTidy(data = NULL)
}
\arguments{
\item{data}{Data to potentially convert.}
}
\value{
'Tidy' formatted `ws_monitor` data.
}
\description{
Changes write-optomized `ws_monitor` formatted data into a read-optomized
'tidy' format that is preferred for `ggplot` plotting. If the given data is
already in a tidy format, it is returned as is.
}
\examples{

}
