% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_toTidy.R
\name{monitor_toTidy}
\alias{monitor_toTidy}
\title{Convert mts_monitor data to a tidy format}
\usage{
monitor_toTidy(monitor = NULL)
}
\arguments{
\item{monitor}{\emph{mts_monitor} object.}
}
\value{
Tidy formatted \emph{mts_monitor} data.
}
\description{
Changes \emph{mts_monitor} formatted data into a
'mts_tidy' format that is useful for 'tidyverse' functions. If the given data
is already in tidy format, it is returned as is.
}
\examples{
library(AirMonitorPlots)

monitor <-
  AirMonitor::NW_Megafires \%>\%
  AirMonitor::monitor_select(
    c('450d08fb5a3e4ea0_530470009', '40ffdacb421a5ee6_530470010')
  )

mts_tidy <- monitor_toTidy(monitor)

}
