% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_stat_nowcast.R
\name{stat_nowcast}
\alias{stat_nowcast}
\title{Add NowCast values to a plot}
\usage{
stat_nowcast(mapping = NULL, data = NULL, version = "pm",
  includeShortTerm = FALSE, geom = "path", position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE,
  timeseries.legend = FALSE, legend.label = "NowCast", ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified and 
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options: 
if \code{NULL}, the default, the data is inherited from the plot data. 
A \code{data.frame} or other object, will override the plot data. 
A \code{function} will be called witha  single argument, the plot data.
The return value must be a \code{data.frame}, and will be used as the layer data.}

\item{version}{character identity specifying the type of nowcast algorithm to be used. 
For details see \link{monitor_nowcast}.}

\item{includeShortTerm}{calculate preliminary NowCast values starting with the 2nd hour.}

\item{geom}{The geometic object to display the data}

\item{position}{Position adjustment, either as a string, or the result of a call to a
position adjustment function.}

\item{na.rm}{remove NA values from data}

\item{show.legend}{logical indicating whether this layer should be included in legends.}

\item{inherit.aes}{if \code{FALSE}, overrides the default aesthetics, rather than combining with them. 
This is most useful for helper functions that define both data and the aesthetics and
shouldn't inherit behaviour from the default plot specificatino, eg \code{borders()}.}

\item{timeseries.legend}{Logical indicating whether to set mappings so that
a legend can later be added using \link{legend_pm25Timeseries}.}

\item{legend.label}{Label for pm25Points part of the legend. Must match 
legend.labels argument in \link{legend_pm25Timeseries}.}

\item{...}{additional arguments passed on to \code{layer()}, such as aesthetics.}
}
\description{
This function calculates the NowCast values for the data, and adds  it to a plot.
The default is to add a NowCast line.
}
