% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{wsMonToTidy}
\alias{wsMonToTidy}
\title{Convert `ws_monitor` Data to a 'Tidy' Format}
\usage{
wsMonToTidy(data = NULL)
}
\arguments{
\item{data}{Data to convert.}
}
\value{
'Tidy' formatted `ws_monitor` data.
}
\description{
Changes write-optomized `ws_monitor` formatted data into a
    read-optomized 'tidy' format.
}
\examples{

}
